fix: change return type of property

+++ src/main/java/com/auction/auto_auction/repository/projection/CarMarkToBid.java
@@ -1,6 +1,8 @@
 package com.auction.auto_auction.repository.projection;
 
+import java.math.BigDecimal;
+
 public interface CarMarkToBid {
     String getCarMark();
-    Long getBidValue();
+    BigDecimal getBidValue();
 }
